Split out the servicex file
The code doesn't need to talk that heavily.
